cql3: create_view_statement: fix check for bound variables
The code for checking that an MV's select statement doesn't
have any bind markers uses the wrong method and always returns
`false` even when it should not.

`prepare_context::empty()` is a misleading name because
it doesn't check if the current instance is empty, but creates
an empty  instance wrapped in a `lw_shared_ptr` instead.
Thus, the code in `create_view_statement::announce_migration()`
checks that the pointer is not empty, which is always false.

Use `get_variable_specifications().empty()` to check that the
specifications vector inside the `prepare_context`
instance is not empty.

Tests: unit(dev)

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>