cql3: Ignore regular index in multi-column WHERE
When the WHERE clause contains a multi-column restriction as well as
an equality on an indexed regular column, we will not use the index.
The index cannot eliminate post-fetch filtering here, but it does add
an extra table query to the execution.

This is similar to db63b4034, which dealt with a clustering-column
index.  Here we deal with a non-key column being indexed.

Fixes #9085.

Tests: unit (dev, debug)

Signed-off-by: Dejan Mircevski <dejan@scylladb.com>