raft: do not let follower's commit_idx to go backwards
append_reply packets can be reordered and thus reply.commit_idx may be
smaller than the one it the tracker. The tracker's commit index is used
to check if a follower needs to be updated with potentially empty
append message, so the bug may theoretically cause unneeded packets to
be sent.

Message-Id: <YQZZ/6nlNb5nQyXp@scylladb.com>