database, storage_proxy: Reconcile pages with no live rows incrementally
Currently, mutation query on replica side will not respond with a
result which doesn't have at least one live row. This causes problems
if there is a lot of dead rows before we reach a live row, which stem
from the fact that resulting reconcilable_result will be large:

1) Large allocations:

  Serialization of reconcilable_result causes large allocations for storing result rows in std::deque

2) Reactor stalls:

  Serialization of reconcilable_reslt on the replica side and on the
  coordinator side causes reactor stalls. This impacts not only the
  query at hand.

  For 1M dead rows, freezing takes 130ms, unfreezing takes 500ms.

  Coordinator does multiple freezes and unfreezes. The reactor stall on the coordinator side is >5s

3) Too large repair mutations

  If reconciliation works on large pages, repair may fail due to too large mutation size.
  1M dead rows is already too much: Refs #9111

This patch fixes all of the above by making mutation reads respect the
memory accounter's limit for the page size, even for dead rows. Range
tombstones are not handled yet, they will be still accumulated (TODO).

The coordinator needs to be prepared for this. Query pagers will
request the next page if it doesn't contain any live row
automatically, like is already done for aggregate queries.

This pach doesn't address the problem of client-side timeouts during
paging. It is an orthogonal problem. Query pager could periodically
send empty pages to the client, or use some other keep-alive
mechanism.

My testing shows that this solution significantly reduces efficiency.
I tested 1M dead rows, which takes 80MB as a single frozen mutation.

  cl=one, no reconciliation, hot cache: 460 ms
  cl=two, reconciling, no paging (before the patch), hot cache: 6.2 s
  cl=two, reconciling, coordinator-side paging (after the patch), hot cache: 24.3 s

The number of page requests in that test was about 80.

Refs #7929 (range tombstone case is not fixed yet)
Refs #3672 (client may still time-out, we should send empty pages)
Refs #7933 (partition tombstone case not fixed yet, see next patch)
           (also, client-side timeout is not solved)
Fixes #9111