raft: fsm: update _commit_idx when applying snapshot
All entries up to snapshot.idx must obviously be committed, so why not
update _commit_idx to reflect that.

With this we get a useful invariant:
`_log.get_snapshot().idx <= _commit_idx`.
For example, when checking whether the latest active configuration is
committed, it should be enough to compare the configuration index to the
commit index. Without the invariant we would need a special case if the
latest configuration comes from a snapshot.