sstables: add handling of reversed clustering ranges
To support multiple clustering ranges in reverse reads, we will
iterate over them in reversed order. Using reversed schema, the
comparisons on clustering keys will get reversed, so we also
switch range starts with ends. For that reason we add methods
for getting reverse starts and ends of a bound which are
different from non-reverse ends and starts in terms of the
bound_kind.