sstables: add support for iterating over clustering ranges in reverse
In the sstable reader, we iterate over clustering ranges using the
index_reader, which normally only accepts advancing to increasing
positions. In this patch we add methods for advancing the index
reader in reverse.

To find the start of the last (first in reverse) row, we may need
to read the entire last promoted index block. To easily find the
position of the last block, a new method for the
bsearch_clustered_cursor is introduced