sstables: add intermediary stream for reversed reads
This patch add the implementation for the input stream, that
returns data buffers with contents of an sstable in such a manner,
as if the rows of the sstable were present in a reversed order.

Because we can't directly communicate between the sstable reader
and the intermediary data source (only the wrapping input_stream),
we're using the index_reader to get the data ranges of the file.
We start with a full partition, and move the upper bound of the
index when we're skipping to the next range.
Currently, skipping between ranges is signaled only by changing
the data range of the index_reader, in contrast to skipping the
input_stream directly in the continuous_data_consumer.
It should be possible to translate the data positions of the
index_reader to the positions of the intermediary data stream,
but I'm not sure we want to do that, as it add some complexity
in the index_reader and the intermediary stream. One reason
for the complexity in the input streamis that we would have two
kinds of skips: a single row skip, and a skip to a clustering
range. We know the offset of the next row, so we could check that
to differentiate them. We would also need to add an information
about the position of first clustering row and end of the last one
in the index_reader.

For simplicity, the intermediary stream always starts with
parsing the partition header and (if present) the static row,
and returning the corresponding bytes as a result of the first
read.
For parsing we're using the same construct as in the sstable parser,
but it only contains a small part of the parsing coroutine and doesn't
perform any correctness checks.

After partition header and static row, we decide whether we need
to read the last promoted index block, (because the data range
ends at the end of the partition), or we can use the previous
unfiltered size from the first row after the range.

After finding the data range of the last row, we parse rows
consecutively in reversed order.
The parser for rows turned out rather big mostly because we can't
always deduce the size of the clustering blocks without reading
the block header.

The parser allows reading rows while skipping their bodies also in
non-reversed order, which we are making use of while reading the
last promoted index block.

The intermediary data source has one more utility, that's currently
completely untested - reversing range tombstones.
When we read a tombstone bound/boundary, we modify the data buffer,
so that the resulting bound/boundary has the reversed kind (so we
don't read ends before starts), and the boundaries have their
before/after timestamps swapped. The end_open_marker isn't
properly considered in this patch either.