sstables: add temporary solution for integrating the reversed sstable reader
To test the sstable reader without having to implement the reversed
memtable/cache readers, use a reversing reader on the new sstable
reader. The resulting reader is reversing the order twice, so
it gives the same results as the non-modified reader, so it's
compatible with the current implementation of reversed reads.