database: Annotate query result with the last primary key on short read
The code touched by this patch is executed by the coordinator during
read-repair when converting reconcilable_result to a data-only query
result.

After this patch, query::result is annotated with the last primary key
present in the reconcilable result. This will be used later by paging
layer to calculate the primary key of the next page. Note that this
information was not present in the query::result before, because it
had dead fragments compacted away. In particular, query::result could
be empty, and yet have a short-read primary key in the middle of
partition.